some more cleaning
